# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Currency Tracker Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "local server"
        }
    ],
    "paths": {
        "/currencies": {
            "get": {
                "tags": [
                    "Currency"
                ],
                "summary": "List of available and active currencies in the system.",
                "operationId": "1158e1c13797931ee1bc6ddef53fc8d7",
                "responses": {
                    "200": {
                        "description": "Success 200",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CurrencyResource"
                                            }
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions/currency/{currency}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "represents the average rate for each currency",
                "operationId": "1efe2a5c05172a7c702a5def1b17c179",
                "parameters": [
                    {
                        "name": "currency",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CurrencyResource"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions/store": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "To record currency receipt, which receives currency type, currency amount form user",
                "operationId": "2f4fd95fc9082822e83fd29afaebef35",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TransactionResource"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "It lists receipts, including which currency,with rate",
                "operationId": "f8c5b9688ff5f335d026e800937e284f",
                "responses": {
                    "200": {
                        "description": "Success auth",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TransactionResource"
                                            }
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreTransactionRequest": {
                "title": "Request store transaction",
                "properties": {
                    "amount": {
                        "type": "string"
                    },
                    "exchange_rate": {
                        "type": "string"
                    },
                    "currency_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CurrencyResource": {
                "title": "Info currency resource",
                "required": [
                    "id",
                    "name",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string"
                    },
                    "createdAt": {
                        "description": "Date and time created comment as ISO format",
                        "type": "string",
                        "example": "2023-04-29 19:00:58"
                    },
                    "updatedAt": {
                        "description": "Date and time updated comment as ISO format",
                        "type": "string",
                        "example": "2023-04-29 19:00:58"
                    }
                },
                "type": "object"
            },
            "TransactionResource": {
                "title": "Info currency resource",
                "required": [
                    "id",
                    "amount",
                    "exchangeRate",
                    "currency",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "type": "string"
                    },
                    "exchangeRate": {
                        "type": "string"
                    },
                    "currency": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "createdAt": {
                        "description": "Date and time created comment as ISO format",
                        "type": "string",
                        "example": "2023-04-29 19:00:58"
                    },
                    "updatedAt": {
                        "description": "Date and time updated comment as ISO format",
                        "type": "string",
                        "example": "2023-04-29 19:00:58"
                    }
                },
                "type": "object"
            }
        }
    }
}